Update Euler-Burger RS with LLF RS
Not working yet. I am passing the space as the first auxiliary array
and the point in space where we want to switch RS as the second
auxiliary array
but it is not working yet.
Probably the best way to go is to use the Burgers RS over the whole
domain to make sure it behaves well first. I am using Local Lax
Friedrich's method for Burger's equation.