Add slowing method, update swithching
Three different solvers are used. An space dependent auxiliary variable
is passed to each RS to determine where to switch between euler
equations to burgers (or to start slowing down waves).

The slowing down of waves is not working properly yet:
I think that the main issue is that the waves are still reaching the
boundary. I need to make the slowing process:
- Increase smoothly, not in a single jump (in the worst case scenario
  several small jump).
- Become more agressive as the waves approach the right boundary.