Add absorbing BC to dissipate mass
After the waves are slowed down, it seems like mass is accumulate in the
last cells, so we require a way to artificially dissipate it. Using a
damping technique like the one proposed by Smadar Karni (at least as it
is) doesn't work, since if ,e.g., V=1/rho is damped, the problem is just
worse. It doesn't suffice either to just damp the velocity, since it
doesn't prevent mass accumulation. What I am doing is to use some
absorbing boundary conditions where the solution is forced to coincide
with some target one. This target solution is chosen to be the constant
initial steady state. It seems to work for the moment, but some open
questions remain:

- How well does it work against the simple wave approach?
- Isn't it better to just use the absorbing BC? (First comparissons not
  so clear)
- Maybe we can dissipate mass in a much better way. For example, solving
 a modified system of equations where damping all the conserved
 quantities lead us to the initial constant steady state. Then it makes
 sense to use an already developed damping technique, like Karni's
 damping of certain waves.