Add modified solver for perturbed system
In order to damp waves to zero using "conventional" damping boundary
conditions, we can modify our system. We just need to translate the
solution to the Euler equations and we get almost the same conservation
law. From the solution to this modified system we can retrieve the
solution to the original system and the trivial solution for this system
will correspond to the constant steady state we used for the
modification.